+++ include/farm.h
@@ -21,10 +21,10 @@
 #ifndef GRPPI_FARM_H
 #define GRPPI_FARM_H
 
-#include "common/sequential_execution.h"
-#include "common/parallel_execution_native.h"
-#include "common/parallel_execution_omp.h"
-#include "common/parallel_execution_tbb.h"
+#include "seq/sequential_execution.h"
+#include "native/parallel_execution_native.h"
+#include "omp/parallel_execution_omp.h"
+#include "tbb/parallel_execution_tbb.h"
 #include "common/support.h"
 
 #include "common/patterns.h"
