+++ include/poly/divideconquer.h
@@ -21,7 +21,7 @@
 #ifndef GRPPI_POLY_DIVIDECONQUER_H
 #define GRPPI_POLY_DIVIDECONQUER_H
 
-#include "common/polymorphic_execution.h"
+#include "polymorphic_execution.h"
 #include "common/support.h"
 
 namespace grppi{
