+++ include/poly/farm.h
@@ -22,7 +22,7 @@
 #ifndef GRPPI_POLY_FARM_H
 #define GRPPI_POLY_FARM_H
 
-#include "common/polymorphic_execution.h"
+#include "polymorphic_execution.h"
 #include "common/support.h"
 
 namespace grppi{
