+++ include/poly/polymorphic_execution.h
@@ -24,10 +24,10 @@
 #include <typeinfo>
 #include <memory>
 
-#include "common/sequential_execution.h"
-#include "common/parallel_execution_native.h"
-#include "common/parallel_execution_tbb.h"
-#include "common/parallel_execution_omp.h"
+#include "seq/sequential_execution.h"
+#include "native/parallel_execution_native.h"
+#include "tbb/parallel_execution_tbb.h"
+#include "omp/parallel_execution_omp.h"
 
 namespace grppi{
