+++ include/poly/stream_iteration.h
@@ -21,7 +21,7 @@
 #ifndef GRPPI_POLY_STREAM_ITERATOR_H
 #define GRPPI_POLY_STREAM_ITERATOR_H
 
-#include "common/polymorphic_execution.h"
+#include "polymorphic_execution.h"
 #include "common/support.h"
 
 namespace grppi{
