+++ unit_tests/mapreduce.cpp
@@ -23,7 +23,7 @@
 #include <iostream>
 
 #include "mapreduce.h"
-#include "common/polymorphic_execution.h"
+#include "poly/polymorphic_execution.h"
 
 #include "supported_executions.h"
