+++ unit_tests/pipeline.cpp
@@ -27,7 +27,7 @@
 #include "farm.h"
 
 #include "supported_executions.h"
-#include "common/polymorphic_execution.h"
+#include "poly/polymorphic_execution.h"
 
 #include <iostream>
 #include <numeric>
