+++ unit_tests/stencil.cpp
@@ -23,7 +23,7 @@
 #include <gtest/gtest.h>
 
 #include "stencil.h"
-#include "common/polymorphic_execution.h"
+#include "poly/polymorphic_execution.h"
 
 #include "supported_executions.h"
