+++ unit_tests/stream_filter.cpp
@@ -24,7 +24,7 @@
 
 #include "stream_filter.h"
 #include "pipeline.h"
-#include "common/polymorphic_execution.h"
+#include "poly/polymorphic_execution.h"
 
 #include "supported_executions.h"
