+++ unit_tests/stream_reduce.cpp
@@ -23,7 +23,7 @@
 #include <gtest/gtest.h>
 
 #include "stream_reduce.h"
-#include "common/polymorphic_execution.h"
+#include "poly/polymorphic_execution.h"
 
 #include "supported_executions.h"
