+++ unit_tests/supported_executions.h
@@ -3,10 +3,10 @@
 
 #include <gtest/gtest.h>
 
-#include "common/sequential_execution.h"
-#include "common/parallel_execution_native.h"
-#include "common/parallel_execution_omp.h"
-#include "common/parallel_execution_tbb.h"
+#include "seq/sequential_execution.h"
+#include "native/parallel_execution_native.h"
+#include "omp/parallel_execution_omp.h"
+#include "tbb/parallel_execution_tbb.h"
 
 using executions = ::testing::Types<
   grppi::sequential_execution,
